fix: add forward-compatibility for new Beet versions (#26)
* chore: use >= constrant for Beet library versions

Dependencies are still locked at their current version, but library
consumers now have the flexibility to use higher `0.x` releases of
Beet and related dependencies.

Additionally `poetry.lock` uses the new format for Poetry 1.3.x, but
1.2.2 used by the GitHub Action is forwards-compatible with it so it
should be fine.

* fix: add missing `run` in some execute commands

The pinned version of Mecha seems to allow `run` to be omitted, but in
new versions of Mecha this is not allowed and causes `beet build` to
fail. Now the build should work with newer versions of Mecha.